+++ src/features/textEditor/components/tipTapEditorConfigs.ts
@@ -9,14 +9,14 @@ import markdown from 'highlight.js/lib/languages/markdown';
 import { lowlight } from 'lowlight';
 import { Markdown } from 'tiptap-markdown';
 
-import { CitationNode } from './tipTapNodes/citation/CitationNode';
+import { CitationNode } from './tipTapNodes/citation/CitationNode2';
 import { CollapsibleBlockContentNode } from './tipTapNodes/collapsibleBlock/nodes/CollapsibleBlockContent';
 import { CollapsibleBlockNode } from './tipTapNodes/collapsibleBlock/nodes/CollapsibleBlockNode';
 import { CollapsibleBlockSummaryNode } from './tipTapNodes/collapsibleBlock/nodes/CollapsibleBlockSummary';
 import { DraggableBlockNode } from './tipTapNodes/draggableBlock/DraggableBlockNode';
-import { ReferenceNode } from './tipTapNodes/references/ReferenceNode';
+import { ReferenceNode } from './tipTapNodes/references/ReferenceNode2';
 import { RefStudioDocument } from './tipTapNodes/refStudioDocument/RefStudioDocument';
-import { SentenceCompletionExtension } from './tipTapNodes/sentenceCompletion/SentenceCompletion';
+import { SentenceCompletionExtension } from './tipTapNodes/sentenceCompletion/SentenceCompletion2';
 lowlight.registerLanguage('markdown', markdown);
 
 export const EDITOR_EXTENSIONS: Extensions = [
