+++ src/features/textEditor/components/tipTapNodes/citation/inputRuleHandlers/squareBracketHandler.ts
@@ -1,7 +1,7 @@
 import { TextSelection } from '@tiptap/pm/state';
 import { InputRule } from '@tiptap/react';
 
-import { CitationNode } from '../CitationNode';
+import { CitationNode } from '../CitationNode2';
 
 export const squareBracketHandler = ({
   state,
